[test] move transformation tests from move_test to transformation_test
Test case was in the wrong file.
